refactor: changed paths
